+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Provide better error output on initial MQTT connection errors
+- Show total amount of messages in the topic overview title
 
 ## [0.20.0] - 2024-02-26